+++ rust/src/models/surf_constants.rs
@@ -1,5 +1,5 @@
 pub struct SurfConstants {
     pub start_hour: u32,
     pub stop_hour: u32,
-    pub wind_speed: f64
+    pub wind_speed: f32
 }
